+++ components/compass/icon/icon.tsx
@@ -2,15 +2,25 @@
 // See LICENSE.txt for license information.
 
 import React from 'react';
+import classNames from 'classnames';
 
-import './icon.scss';
-
-export default class Icon extends React.PureComponent {
-    render() {
-        return (
-            <i className={`Icon Icon___standard ${iconGlyphs.EMOTICON_HAPPY_OUTLINE}`}/>);
-    }
+interface Props {
+    glyph: string;
+    size: string;
+    label: string;
+    additionalClass?: string;
 }
+
+const Icon: React.FC<Props> = ({glyph, size, label, additionalClass}: Props) => {
+    return (
+        <i
+            className={classNames([`Icon Icon___${size} icon-${glyph}`, {additionalClass}])}
+            aria-label={label}
+        />
+    );
+};
+
+export default Icon;
 
 export const iconGlyphs = {
     EMOTICON_HAPPY_OUTLINE: 'icon-emoticon-happy-outline',
